rename file to primitive
