+++ rendering/src/lib.rs
@@ -1,7 +1,7 @@
-pub mod render_primitive;
+pub mod primitive;
 pub mod renderable;
 pub mod style;
 pub mod svg_plot;
 
-pub use render_primitive::{CircularArc, CircularArcTo, PathCommand, RenderPrimitive};
+pub use primitive::{CircularArc, CircularArcTo, PathCommand, RenderPrimitive};
 pub use renderable::Renderable;
