+++ rendering/src/primitive.rs
