+++ rendering/src/renderable.rs
@@ -1,6 +1,6 @@
 use std::error::Error;
 
-use crate::render_primitive::RenderPrimitive;
+use crate::primitive::RenderPrimitive;
 
 pub trait Renderable {
     fn bake_geometry(&self) -> Result<Vec<RenderPrimitive>, Box<dyn Error>>;
